refactor(tracker-core): split database drivers into folder modules
