+++ packages/tracker-core/src/databases/driver/sqlite/auth_key_store.rs
@@ -1,0 +1,105 @@
+use std::panic::Location;
+
+use r2d2_sqlite::rusqlite::params;
+use r2d2_sqlite::rusqlite::types::Null;
+use torrust_tracker_primitives::DurationSinceUnixEpoch;
+
+use super::{Sqlite, DRIVER};
+use crate::authentication::{self, Key};
+use crate::databases::error::Error;
+use crate::databases::AuthKeyStore;
+
+impl AuthKeyStore for Sqlite {
+    fn load_keys(&self) -> Result<Vec<authentication::PeerKey>, Error> {
+        let conn = self.pool.get().map_err(|e| (e, DRIVER))?;
+
+        let mut stmt = conn.prepare("SELECT key, valid_until FROM keys")?;
+
+        let keys_iter = stmt.query_map([], |row| {
+            let key: String = row.get(0)?;
+            let opt_valid_until: Option<i64> = row.get(1)?;
+
+            match opt_valid_until {
+                Some(valid_until) => Ok(authentication::PeerKey {
+                    key: key.parse::<Key>().unwrap(),
+                    valid_until: Some(DurationSinceUnixEpoch::from_secs(valid_until.unsigned_abs())),
+                }),
+                None => Ok(authentication::PeerKey {
+                    key: key.parse::<Key>().unwrap(),
+                    valid_until: None,
+                }),
+            }
+        })?;
+
+        let keys: Vec<authentication::PeerKey> = keys_iter.filter_map(std::result::Result::ok).collect();
+
+        Ok(keys)
+    }
+
+    fn get_key_from_keys(&self, key: &Key) -> Result<Option<authentication::PeerKey>, Error> {
+        let conn = self.pool.get().map_err(|e| (e, DRIVER))?;
+
+        let mut stmt = conn.prepare("SELECT key, valid_until FROM keys WHERE key = ?")?;
+
+        let mut rows = stmt.query([key.to_string()])?;
+
+        let key = rows.next()?;
+
+        Ok(key.map(|f| {
+            let valid_until: Option<i64> = f.get(1).unwrap();
+            let key: String = f.get(0).unwrap();
+
+            match valid_until {
+                Some(valid_until) => authentication::PeerKey {
+                    key: key.parse::<Key>().unwrap(),
+                    valid_until: Some(DurationSinceUnixEpoch::from_secs(valid_until.unsigned_abs())),
+                },
+                None => authentication::PeerKey {
+                    key: key.parse::<Key>().unwrap(),
+                    valid_until: None,
+                },
+            }
+        }))
+    }
+
+    fn add_key_to_keys(&self, auth_key: &authentication::PeerKey) -> Result<usize, Error> {
+        let conn = self.pool.get().map_err(|e| (e, DRIVER))?;
+
+        let insert = match auth_key.valid_until {
+            Some(valid_until) => conn.execute(
+                "INSERT INTO keys (key, valid_until) VALUES (?1, ?2)",
+                [auth_key.key.to_string(), valid_until.as_secs().to_string()],
+            )?,
+            None => conn.execute(
+                "INSERT INTO keys (key, valid_until) VALUES (?1, ?2)",
+                params![auth_key.key.to_string(), Null],
+            )?,
+        };
+
+        if insert == 0 {
+            Err(Error::InsertFailed {
+                location: Location::caller(),
+                driver: DRIVER,
+            })
+        } else {
+            Ok(insert)
+        }
+    }
+
+    fn remove_key_from_keys(&self, key: &Key) -> Result<usize, Error> {
+        let conn = self.pool.get().map_err(|e| (e, DRIVER))?;
+
+        let deleted = conn.execute("DELETE FROM keys WHERE key = ?", [key.to_string()])?;
+
+        if deleted == 1 {
+            // should only remove a single record.
+            Ok(deleted)
+        } else {
+            Err(Error::DeleteFailed {
+                location: Location::caller(),
+                error_code: deleted,
+                driver: DRIVER,
+            })
+        }
+    }
+}
