+++ packages/tracker-core/src/databases/driver/sqlite/schema_migrator.rs
@@ -1,0 +1,69 @@
+use super::{Sqlite, DRIVER};
+use crate::databases::error::Error;
+use crate::databases::SchemaMigrator;
+
+impl SchemaMigrator for Sqlite {
+    fn create_database_tables(&self) -> Result<(), Error> {
+        let create_whitelist_table = "
+        CREATE TABLE IF NOT EXISTS whitelist (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            info_hash TEXT NOT NULL UNIQUE
+        );"
+        .to_string();
+
+        let create_torrents_table = "
+        CREATE TABLE IF NOT EXISTS torrents (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            info_hash TEXT NOT NULL UNIQUE,
+            completed INTEGER DEFAULT 0 NOT NULL
+        );"
+        .to_string();
+
+        let create_torrent_aggregate_metrics_table = "
+        CREATE TABLE IF NOT EXISTS torrent_aggregate_metrics (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            metric_name TEXT NOT NULL UNIQUE,
+            value INTEGER DEFAULT 0 NOT NULL
+        );"
+        .to_string();
+
+        let create_keys_table = "
+        CREATE TABLE IF NOT EXISTS keys (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            key TEXT NOT NULL UNIQUE,
+            valid_until INTEGER
+         );"
+        .to_string();
+
+        let conn = self.pool.get().map_err(|e| (e, DRIVER))?;
+
+        conn.execute(&create_whitelist_table, [])?;
+        conn.execute(&create_keys_table, [])?;
+        conn.execute(&create_torrents_table, [])?;
+        conn.execute(&create_torrent_aggregate_metrics_table, [])?;
+
+        Ok(())
+    }
+
+    fn drop_database_tables(&self) -> Result<(), Error> {
+        let drop_whitelist_table = "
+        DROP TABLE whitelist;"
+            .to_string();
+
+        let drop_torrents_table = "
+        DROP TABLE torrents;"
+            .to_string();
+
+        let drop_keys_table = "
+        DROP TABLE keys;"
+            .to_string();
+
+        let conn = self.pool.get().map_err(|e| (e, DRIVER))?;
+
+        conn.execute(&drop_whitelist_table, [])
+            .and_then(|_| conn.execute(&drop_torrents_table, []))
+            .and_then(|_| conn.execute(&drop_keys_table, []))?;
+
+        Ok(())
+    }
+}
